requirements_cluster.txt
absl-py==2.1.0
alembic==1.14.1
astunparse==1.6.3
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chex==0.1.88
cloudpickle==1.6.0
colorlog==6.9.0
contourpy==1.3.1
cycler==0.12.1
cython==3.1.0
DataProperty==1.1.0
decorator==5.1.1
dm-tree==0.1.8
dotmap==1.3.30
etils==1.11.0
evosax==0.1.6
Farama-Notifications==0.0.4
fasteners==0.19
filelock==3.16.1
flatbuffers==24.12.23
flax==0.10.2
fonttools==4.55.3
fsspec==2024.12.0
gast==0.6.0
glfw==2.8.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.69.0
gym-notices==0.0.8
gymnasium==0.29.0
gymnax==0.0.8
h5py==3.12.1
huggingface-hub==0.27.1
huggingface-sb3==3.0
humanize==4.11.0
idna==3.10
imageio==2.37.0
importlib-resources==6.5.2
jax==0.5.0
jax-cuda12-pjrt==0.5.0
jax-cuda12-plugin==0.5.0
jaxlib==0.5.0
jinja2==3.1.5
joblib==1.4.2
keras==3.8.0
kiwisolver==1.4.8
libclang==18.1.1
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mbstrdecoder==1.1.4
mdurl==0.1.2
ml-collections==1.0.0
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.0
mujoco==3.3.1
mujoco-mjx==3.3.1
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvcc-cu12==12.6.85
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
opt-einsum==3.4.0
optax==0.2.4
optree==0.14.0
optuna==4.2.0
orbax-checkpoint==0.11.0
packaging==24.2
pandas==2.2.3
pathvalidate==3.2.3
pettingzoo==1.24.3
pillow==11.1.0
protobuf==5.29.3
pycparser==2.22
pyglet==1.5.27
pygments==2.19.1
PyOpenGL==3.1.9
pyparsing==3.2.1
pytablewriter==1.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
sb3-contrib==2.3.0
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
Shimmy==1.3.0
simplejson==3.19.3
six==1.17.0
SQLAlchemy==2.0.37
stable-baselines3==2.3.2
sympy==1.13.1
tabledata==1.3.4
tcolorpy==0.1.7
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-probability==0.25.0
tensorstore==0.1.71
termcolor==2.5.0
tf-keras==2.18.0
threadpoolctl==3.5.0
toolz==1.0.0
torch==2.5.1
tqdm==4.67.1
trimesh==4.6.8
triton==3.1.0
typepy==1.3.4
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wasabi==1.1.3
werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
